requirements.txt
langchain-google-genai
langchain-community
langgraph
python-dotenv
fastapi
uvicorn
protobuf==4.23.4
